update: ゲーム開始時処理のインターフェースを追加
