+++ Assets/Prefabs/RunTimeSystem/InGameManager.prefab
@@ -100,6 +100,8 @@ MonoBehaviour:
   _cameraOffset: {x: 0, y: 1, z: -3}
   _setIcon: {fileID: 0}
   _isRandomSpawn: 0
+  _gameStartStrategy:
+    rid: 7765555065531400345
   _gameStartPerformances:
   - rid: 7765555050530472034
   - rid: 7765555050530472041
@@ -121,6 +123,9 @@ MonoBehaviour:
         _enabled: 1
     - rid: 7765555065531400310
       type: {class: TagPlayerKilledUseCase, ns: September.Common, asm: Scripts}
+      data: 
+    - rid: 7765555065531400345
+      type: {class: TagGameStartStrategy, ns: September.Common, asm: Scripts}
       data: 
 --- !u!114 &4864244944479918103
 MonoBehaviour:
